moved about/darkmode to footer, fix footer on mobile
